feat: add ScreenController function
add a reference to, and expose, the getBoard
method within GameController to help maintain
encapsulation while allowing ScreenController
access.

add ScreenController and the logic to render and
update the screen after each turn.

generate the board as buttons on the screen, and
display the current active player.

add defer tag to html script tag to ensure the
script runs after the DOM loads.