feat: add win/tie modal w/ refactoring
have a modal appear on a win or tie with buttons
allowing the players to play again with the same
names, or go back to the initial home screen where
they can pick new names and start again.

adjust html to have the modal be filled with
content dynamically from the DOM.

add showModal, hideModal, bindModalEvents,
homeScreenModal, and endGameModal functions to
ScreenController.

refactor JS to make this functionality work.

style modal and buttons with CSS.